tests: factorize given_call_not_stasis
